+++ __tests__/commands/changelogs/single.test.ts
@@ -6,7 +6,7 @@ import frontMatter from 'gray-matter';
 import nock from 'nock';
 import { describe, beforeAll, afterAll, beforeEach, it, expect } from 'vitest';
 
-import Command from '../../../src/cmds/changelogs.js';
+import Command from '../../../src/commands/changelogs.js';
 import APIError from '../../../src/lib/apiError.js';
 import getAPIMock from '../../helpers/get-api-mock.js';
 import hashFileContents from '../../helpers/hash-file-contents.js';
