+++ __tests__/commands/open.test.ts
@@ -1,10 +1,10 @@
-import type { Version } from '../../src/cmds/versions/index.js';
+import type { Version } from '../../src/commands/versions/index.js';
 
 import chalk from 'chalk';
 import { describe, afterEach, beforeAll, it, expect } from 'vitest';
 
 import pkg from '../../package.json';
-import Command from '../../src/cmds/open.js';
+import Command from '../../src/commands/open.js';
 import configStore from '../../src/lib/configstore.js';
 import getAPIMock from '../helpers/get-api-mock.js';
 import { runCommand } from '../helpers/setup-oclif-config.js';
