+++ package.json
@@ -178,7 +178,7 @@
         "description": "Sync or prune Guides pages in your ReadMe developer hub."
       },
       "openapi": {
-        "description": "Manage your API definition. Supports OpenAPI, Swagger, Postman."
+        "description": "Manage your API definition. Supports OpenAPI, Swagger, and Postman."
       },
       "versions": {
         "description": "Manage your documentation versions."
