+++ src/commands/categories/create.ts
