+++ src/commands/docs/index.ts
