+++ src/commands/open.ts
