+++ src/commands/openapi/inspect.ts
