+++ src/commands/versions/delete.ts
