+++ src/lib/versionSelect.ts
@@ -1,5 +1,5 @@
 import type { APIErrorResponse } from './apiError.js';
-import type { Version } from '../cmds/versions/index.js';
+import type { Version } from '../commands/versions/index.js';
 
 import APIError from './apiError.js';
 import isCI from './isCI.js';
